+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Change the default span name to be `GET /path` so it complies with the HTTP semantic conventions in `go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux`. (#7551)
+- Chang the default By-unit explicit histogram bucket bounds to [0, 50MB) for `google.golang.org/grpc` and `net/http` instrumentation packages. (#7696)
 
 ### Deprecated